+++ src/components/NewsletterSignup/newsletterSignupChangelogData.ts
@@ -19,6 +19,15 @@ export const changelogData: ChangelogData[] = [
     ],
   },
   {
+    date: "2023-11-09",
+    version: "2.1.2",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      "Updates the dark mode color variants for the `newsletterSignupType` prop.",
+    ],
+  },
+  {
     date: "2023-10-26",
     version: "2.1.1",
     type: "Update",