+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 2.1.2 (November 9, 2023)
+
 ### Adds
 
 - Adds the `useDSHeading` hook to render a default H2 heading or a custom heading element.
@@ -20,6 +22,13 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the following components to accept `JSX.Element` type values into their `"headingText"` prop: `AlphabetFilter`, `AudioPlayer`, `ComponentWrapper`, `SearchBar`, `VideoPlayer`
 - Updates the `Notification` component to accept `JSX.Element` type values into its `"notificationHeading"` prop.
 - Updates the `StructuredContent` component to accept `JSX.Element` type values into its `"headingText"` and `"calloutText"` props.
+
+### Updates
+
+- Updates max version for Chakra UI to `1.8.9` for minor bug fixes.
+- Updates the layout for the `"campaign"` variant of the `Hero` component to have consistent padding on its left and right sides.
+- Updates the `getSectionColors` function to also return dark mode color variants.
+- Updates the `NewsletterSignup` component's `newsletterSignupType` prop to render proper dark mode variant colors.
 
 ## 2.1.1 (October 26, 2023)